refactor: move logic to retrieve data for projects to container
