+++ src/pages/projects.tsx
@@ -4,7 +4,6 @@ import Head from '../components/Head';
 import Layout from '../components/Layout';
 import Stars from '../components/Stars';
 import Content from '../components/Content';
-import {Work} from '../types';
 import Projects from '../components/Projects';
 
 interface ProjectsPageProps {
@@ -13,9 +12,7 @@ interface ProjectsPageProps {
 
 const ProjectsPage: React.FC<ProjectsPageProps> = ({data}) => {
   const siteTitle = data.site.siteMetadata.title;
-  const projects: Work[] = data.allMarkdownRemark.edges.filter(
-    work => work.node.frontmatter.type === 'project'
-  );
+
   return (
     <Layout title={siteTitle}>
       <Head
@@ -35,7 +32,7 @@ const ProjectsPage: React.FC<ProjectsPageProps> = ({data}) => {
         addEventListeners={false}
       />
       <Content>
-        <Projects projects={projects} />
+        <Projects />
       </Content>
     </Layout>
   );
@@ -47,26 +44,6 @@ interface PageQueryData {
       title: string;
     };
   };
-  allMarkdownRemark: {
-    edges: {
-      node: {
-        id: string;
-        excerpt: string;
-        fields: {
-          slug: string;
-        };
-        frontmatter: {
-          date: string;
-          title: string;
-          description: string;
-          tags: string[];
-          type: string;
-          url: string;
-          featuredImage: any;
-        };
-      };
-    }[];
-  };
 }
 
 export const pageQuery = graphql`
@@ -74,35 +51,6 @@ export const pageQuery = graphql`
     site {
       siteMetadata {
         title
-      }
-    }
-    allMarkdownRemark(
-      filter: {frontmatter: {published: {ne: false}}}
-      sort: {fields: [frontmatter___date], order: DESC}
-    ) {
-      edges {
-        node {
-          id
-          excerpt
-          fields {
-            slug
-          }
-          frontmatter {
-            date(formatString: "MMM DD, YYYY")
-            title
-            description
-            tags
-            type
-            url
-            featuredImage {
-              childImageSharp {
-                fluid(maxWidth: 800) {
-                  ...GatsbyImageSharpFluid
-                }
-              }
-            }
-          }
-        }
       }
     }
   }
